Convert charset

+++ local_utl/char2hex/char2hex.c
@@ -2,11 +2,11 @@
 #include <string.h>
 
 int main(int argc, char *argv[]) {
-	int i, l;
+	unsigned long i, l;
 	unsigned long j;
 
-	for (i=1; i<argc; i++) {
-		for (j=0, l=strlen(argv[i]); j<l; j++) {
+	for (i = 1; i < (unsigned long) argc; i++) {
+		for (j = 0, l = strlen(argv[i]); j < l; j++) {
 			printf("\\x%X", argv[i][j] & 0xff);
 		}
 		printf("\n");
